pyproject.toml
[tool.poetry]
name = "mdbl"
version = "0.1.0"
description = ""
packages = [{ include = "mdbl", from = "src" }]
authors = ["hectorxh <hectorplj@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mdbl = "mdbl.cli:main"
mdbl_dev = "mdbl.cli:dev_utils"

[tool.poetry.dependencies]
python = "^3.12"
duckdb = "^1.1.0"
click = "^8.1.7"
pydantic = "^2.9.1"
faker = "^28.4.1"
polars = "^1.7.1"
pyyaml = "^6.0.2"
psycopg = { extras = ["binary"], version = "^3.2.2" }

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportAny = false
reportUnknownMemberType = false
reportUnusedCallResult = false

[tool.ruff]
target-version = "py312"
